% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess.lasso.ridge}
\alias{preprocess.lasso.ridge}
\title{Preprocess input data with LASSO/Ridge regregression method}
\usage{
preprocess.lasso.ridge(data, y, reg.family, method)
}
\arguments{
\item{data}{An input matrix with values of independent variables (predictors).}

\item{y}{A vector with values of dependent variable (outcome).}

\item{reg.family}{A regression family. Can be either "binomial" or "gaussian."}

\item{method}{A method. Can be either "lasso" or "ridge."}
}
\value{
fit An object returned by "cv.glmnet" function.
}
\description{
Preprocess input data with LASSO/Ridge regregression method
}

