% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess.plsda}
\alias{preprocess.plsda}
\title{Preprocess input data with Partial Linear Square 
Regregression Discriminant Analysis method (PLSDA)}
\usage{
preprocess.plsda(data, y, scale = FALSE, verbose = FALSE)
}
\arguments{
\item{data}{An input matrix with values of independent 
variables (predictors).}

\item{y}{A vector with values of dependent variable (outcome).}

\item{scale}{A logical variable, indicates wheither or 
not scaling should be performed.}

\item{verbose}{Indicates verbosing output. Default: FALSE.}
}
\value{
A list of one: "S" - a data frame of predictor values.
}
\description{
Preprocess input data with Partial Linear Square 
Regregression Discriminant Analysis method (PLSDA)
}

