% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{prerocess.pca}
\alias{prerocess.pca}
\title{Preprocess input data with Principal Component Analysis method (PCA)}
\usage{
prerocess.pca(data, scale, cumvar.threshold, verbose = FALSE)
}
\arguments{
\item{data}{An input matrix with values of 
independent variables (predictors).}

\item{scale}{A logical variable, indicates wheither or 
not scaling should be performed.}

\item{cumvar.threshold}{A threshold value for explained variance.}

\item{verbose}{Indicates verbosing output. Default: FALSE.}
}
\value{
A list of one: "S" - a data frame of predictor values.
}
\description{
Preprocess input data with Principal Component Analysis method (PCA)
}

