% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqt.R
\name{QTest.one}
\alias{QTest.one}
\title{This function performs a single (no permutations) gene-level test based on combined effect sizes.}
\usage{
QTest.one(phenotype, genotype, covariates, STT = 0.2, weight = FALSE,
  cumvar.threshold = 90, method = "pca", out.type = "D", scale = FALSE)
}
\arguments{
\item{phenotype}{Phenotype (a matrix n by 1 where n - number of individuals).}

\item{genotype}{Genotype (matrix n by m where n - number of individuals, m - number of genetic variants).}

\item{covariates}{A matrix of covariates. Default: NULL.}

\item{STT}{Numeric indicating soft truncation threshold (STT) to convert to gamma parameter (must be <= 0.4). 
Needed for an optimal parameter a in Gamma-distribution. Default: 0.2. 
See, for example, Fridley, et al 2013: "Soft truncation thresholding for gene set analysis of RNA-seq data: Application to a vaccine study".}

\item{weight}{Logical value. Indicates using weights (see Lee et al 2016). Default: FALSE.}

\item{cumvar.threshold}{Numeric value indicating the explained variance threshold for PCA-like methods. Default: 90.}

\item{method}{Method used to reduce multicollinerity and account for LD. Default: PCA.}

\item{out.type}{Character, indicating a type of phenotype. Possible values: D (dichotomous or binary), 
C (continous or qualitative).}

\item{scale}{A logic parameter (TRUE/FALSE) indicating scaling of the genotype dataset.}
}
\value{
A list of two: test statistics (Q1, Q2, Q3), p-values (p1.Q1, p2.Q2, p3.Q3).
}
\description{
This function performs a single (no permutations) gene-level test based on combined effect sizes.
}

