% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{ridge_se}
\alias{ridge_se}
\title{Calculates a standard error (SE) from glmnet.}
\usage{
ridge_se(xs, y, yhat, my_mod, verbose = FALSE)
}
\arguments{
\item{xs}{A matrix of predictors.}

\item{y}{A vector of outputs (dependent variable).}

\item{yhat}{A vector of predicted values
from \code{glmnet.predict(...)}.}

\item{my_mod}{An object returned from glmnet.}

\item{verbose}{Indicates verbosing output,
FALSE by default.}
}
\value{
A list of two: variance-covariance matrix, 
standard deviations of coefficients.
}
\description{
Calculates a standard error (SE) from glmnet.
}

