% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0classes.R
\docType{methods}
\name{phenotype,rqt-method}
\alias{covariates,rqt-method}
\alias{covariates.rqt}
\alias{genotype,rqt-method}
\alias{genotype.rqt}
\alias{phenotype,rqt-method}
\alias{phenotype.rqt}
\alias{print,rqt-method}
\alias{print.rqt}
\alias{results,rqt-method}
\alias{results.rqt}
\alias{results<-,rqt-method}
\alias{results<-.rqt}
\alias{show,rqt-method}
\alias{show.rqt}
\alias{summary,rqt-method}
\alias{summary.rqt}
\title{Basic methods for class rqt}
\usage{
\S4method{phenotype}{rqt}(x)

\S4method{genotype}{rqt}(x)

\S4method{covariates}{rqt}(x)

\S4method{results}{rqt}(x)

\S4method{results}{rqt}(x) <- value

\S4method{print}{rqt}(x)

\S4method{show}{rqt}(object)

\S4method{summary}{rqt}(object)
}
\arguments{
\item{x}{An object of \code{rqt} class}
}
\value{
pheotype returns the phenotype

genotype returns the genotype

covariates returns the covariates

results returns the results

print returns summary information about the rqt object

show returns summary information about 
the object of class rqt

summary returns the integrated results from the analysis
}
\description{
This document lists a series of basic methods for the class rqt
}

