\name{rqt-package}
\alias{rqt-package}
\alias{rqt}
\docType{package}
\title{
\packageTitle{rqt}
}
\description{
\packageDescription{rqt}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rqt}
\packageIndices{rqt}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rqt}

Maintainer: \packageMaintainer{rqt}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
