% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQTest.R
\docType{methods}
\name{covariates}
\alias{covariates}
\alias{covariates,rqt-method}
\title{This function performs an access to covariates}
\usage{
covariates(x)

\S4method{covariates}{rqt}(x)
}
\arguments{
\item{x}{An object of \code{rqt} class.}
}
\value{
covariates returns the covariates
}
\description{
This function performs an access to covariates

An accessor to covariates
}

