% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQTest.R
\docType{methods}
\name{genotype}
\alias{genotype}
\alias{genotype,rqt-method}
\title{This function performs an access to genotype.}
\usage{
genotype(x)

\S4method{genotype}{rqt}(x)
}
\arguments{
\item{x}{An object of \code{rqt} class.}
}
\value{
genotype returns the genotype
}
\description{
This function performs an access to genotype.

A genotype accessor
}

