% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQTest.R
\docType{methods}
\name{results}
\alias{results}
\alias{results,rqt-method}
\title{This function performs an access to covariates}
\usage{
results(x)

\S4method{results}{rqt}(x)
}
\arguments{
\item{x}{An object of \code{rqt} class.}
}
\value{
results returns the results
}
\description{
This function performs an access to covariates

An accessor to results
}

