% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQTestMeta.R
\docType{methods}
\name{rQTestMeta}
\alias{rQTestMeta}
\alias{rQTestMeta,list-method}
\title{This function performs a gene-level meta-analysis based on 
combined effect sizes.}
\usage{
rQTestMeta(x, ...)

\S4method{rQTestMeta}{list}(x, perm = 0, STT = 0.2, weight = FALSE,
  cumvar.threshold = 90, out.type = "D", method = "pca", scale = FALSE)
}
\arguments{
\item{x}{List of objects of class rqt}

\item{...}{Additional parameters to pass to the function}

\item{perm}{Integer indicating the number of permutations 
to compute p-values. Default: 0.}

\item{STT}{Numeric indicating soft truncation threshold (STT) 
to convert to gamma parameter (must be <= 0.4). 
Needed for an optimal parameter a in Gamma-distribution. Default: 0.2. 
See, for example, Fridley, et al 2013: "Soft truncation thresholding 
for gene set analysis of RNA-seq data: Application to a vaccine study".}

\item{weight}{Logical value. Indicates using weights (see Lee et al 2016). 
Default: FALSE.}

\item{cumvar.threshold}{Numeric value indicating the explained 
variance threshold for PCA-like methods. Default: 90.}

\item{out.type}{Character, indicating a type of phenotype. 
Possible values: D (dichotomous or binary), 
C (continous or qualitative).}

\item{method}{Method used to reduce multicollinerity and account 
for LD. Default: PCA.}

\item{scale}{A logic parameter (TRUE/FALSE) indicating 
scaling of the genotype dataset.}
}
\value{
A list of four: (i) final.pvalue - 
a final p-value across all studies; 
(ii) pvalueList - p-values for each study; 
(iii) df - degrees of freedom; 
(iv) chi.comb - meta-analysis Chi-square statistics
}
\description{
This function performs a gene-level meta-analysis based on 
combined effect sizes.

This function performs a gene-level test based on combined effect sizes.
}
\examples{
data1 <- read.table(system.file("data/phengen2.dat",package="rqt"), skip=1)
obj1 <- rqtClass(phenotype=data1[,1], genotype=data1[, 2:dim(data1)[2]])
data2 <- read.table(system.file("data/phengen3.dat",package="rqt"), skip=1)
obj2 <- rqtClass(phenotype=data2[,1], genotype=data2[, 2:dim(data2)[2]])
data3 <- read.table(system.file("data/phengen.dat",package="rqt"), skip=1)
obj3 <- rqtClass(phenotype=data3[,1], genotype=data3[, 2:dim(data3)[2]])
# Gene-level meta-analysis:
res <- rQTestMeta(list(obj1, obj2, obj3))
print(res)
}

