% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQTestMeta.R
\name{rQTestMeta}
\alias{rQTestMeta}
\title{This function performs a gene-level test based on combined effect sizes.}
\usage{
rQTestMeta(data, covariates, perm = 0, STT = 0.2, weight = FALSE,
  cumvar.threshold = 90, out.type = "D", method = "pca", scale = FALSE)
}
\arguments{
\item{data}{A matrix n by 1 where n - number of individuals).}

\item{covariates}{A matrix of covariates. Default: NULL.}

\item{perm}{Integer indicating the number of permutations to compute p-values. Default: 0.}

\item{STT}{Numeric indicating soft truncation threshold (STT) to convert to gamma parameter (must be <= 0.4). 
Needed for an optimal parameter a in Gamma-distribution. Default: 0.2. 
See, for example, Fridley, et al 2013: "Soft truncation thresholding for gene set analysis of RNA-seq data: Application to a vaccine study".}

\item{weight}{Logical value. Indicates using weights (see Lee et al 2016). Default: FALSE.}

\item{cumvar.threshold}{Numeric value indicating the explained variance threshold for PCA-like methods. Default: 90.}

\item{out.type}{Character, indicating a type of phenotype. Possible values: D (dichotomous or binary), 
C (continous or qualitative).}

\item{method}{Method used to reduce multicollinerity and account for LD. Default: PCA.}

\item{scale}{A logic parameter (TRUE/FALSE) indicating scaling of the genotype dataset.}
}
\value{
A list of two: test statistics (Q1, Q2, Q3), p-values (p1.Q1, p2.Q2, p3.Q3).
}
\examples{

library(rqt)
data1 <- read.table(system.file("data/phengen2.dat",package="rqt"), skip=1)
data2 <- read.table(system.file("data/phengen3.dat",package="rqt"), skip=1)
data3 <- read.table(system.file("data/phengen.dat",package="rqt"), skip=1)
# Combining datasets:
indata <- list(data1, data2, data3)
# Gene-level meta-analysis:
res <- rQTestMeta(indata)
res
}

