% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{simple.multvar.reg}
\alias{simple.multvar.reg}
\title{Applies linear of logistic regregression to the data.}
\usage{
simple.multvar.reg(y, data, reg.family, verbose = FALSE)
}
\arguments{
\item{y}{A vector with values of dependent variable (outcome).}

\item{data}{An input matrix with values of 
independent variables (predictors).}

\item{reg.family}{A regression family. 
Can be either "binomial" or "gaussian."}

\item{verbose}{Indicates verbosing output. Default: FALSE.}
}
\value{
A list of two: "S" - a dataframe with predictors and "fit"
- an object returned by "glm" function.
}
\description{
Applies linear of logistic regregression to the data.
}

