% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{vcov_rigde}
\alias{vcov_rigde}
\title{Calculates variance-covariance matrix for LASSO/ridge regression.}
\usage{
vcov_rigde(x, y, rmod, verbose = FALSE)
}
\arguments{
\item{x}{A matrix of predictors.}

\item{y}{A vector of outputs (dependent variable).}

\item{rmod}{An object returned from glmnet.}

\item{verbose}{Indicates verbosing output. Default: FALSE.}
}
\value{
A list of two: variance-covariance matrix, 
standard deviations of coefficients.
}
\description{
Calculates variance-covariance matrix for LASSO/ridge regression.
}

