% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess.pls}
\alias{preprocess.pls}
\title{Preprocess input data with Partial Linear Square Regregression method (PLS)}
\usage{
preprocess.pls(data, y, cumvar.threshold)
}
\arguments{
\item{data}{An input matrix with values of independent variables (predictors).}

\item{y}{A vector with values of dependent variable (outcome).}

\item{cumvar.threshold}{A threshold value for explained variance.}
}
\value{
A list of one: "S" - a data frame of predictors.
}
\description{
Preprocess input data with Partial Linear Square Regregression method (PLS)
}

