% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqt.R
\name{QTest.one}
\alias{QTest.one}
\title{This function performs a single (no permutations) gene-level test based on combined effect sizes.}
\usage{
QTest.one(y, covadat = NULL, newgeno, STT = 0.2, weight = FALSE,
  cumvar.threshold = 90, method = "pca", out.type = "D")
}
\arguments{
\item{y}{Phenotype (vector).}

\item{covadat}{A matrix of covariates (NULL) by default.}

\item{newgeno}{Genotype (matrix n by m where n - number of individuals, m - number of genetic variants)/}

\item{STT}{TODO}

\item{weight}{Logical. TODO}

\item{cumvar.threshold}{Numeric value indicating the explained variance threshold for PCA-like methods. Default: 90.}

\item{method}{Method used to reduce multicollinerity and account for LD. Default: PCA.}

\item{out.type}{Character, indicating a type of phenotype. Possible values: D (dichotomous or binary), 
C (continous or qualitative).}
}
\value{
A list with p-values and corresponding statistics.
}
\description{
TODO
}

