% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQTest.R
\docType{methods}
\name{genotype}
\alias{genotype}
\alias{genotype,rqt-method}
\title{This function performs an access to genotype.}
\usage{
genotype(x)

\S4method{genotype}{rqt}(x)
}
\arguments{
\item{x}{An object of \code{rqt} class.}
}
\value{
genotype returns the genotype
}
\description{
This function performs an access to genotype.

A genotype accessor
}
\examples{
\dontrun{
data <- data.matrix(read.table(system.file("extdata/test.bin1.dat", package="rqt"), header=TRUE))
pheno <- data[,1]
geno <- data[, 2:dim(data)[2]]
colnames(geno) <- paste(seq(1, dim(geno)[2]))
geno.obj <- SummarizedExperiment(geno)
obj <- rqtClass(phenotype=pheno, genotype=geno.obj)
genotype(obj)
}
}

