% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQTest.R
\docType{methods}
\name{phenotype}
\alias{phenotype}
\alias{phenotype,rqt-method}
\title{This function performs an access to phenotype}
\usage{
phenotype(x)

\S4method{phenotype}{rqt}(x)
}
\arguments{
\item{x}{An object of \code{rqt} class.}
}
\value{
phenotype returns the genotype
}
\description{
This function performs an access to phenotype

A genotype accessor
}

