% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQTest.R
\docType{methods}
\name{results}
\alias{results}
\alias{results,rqt-method}
\title{This function performs an access to covariates}
\usage{
results(x)

\S4method{results}{rqt}(x)
}
\arguments{
\item{x}{An object of \code{rqt} class.}
}
\value{
results returns the results
}
\description{
This function performs an access to covariates

An accessor to results
}
\examples{
data <- data.matrix(read.table(system.file("extdata/test.bin1.dat", 
package="rqt"), header=TRUE))
pheno <- data[,1]
geno <- data[, 2:dim(data)[2]]
colnames(geno) <- paste(seq(1, dim(geno)[2]))
geno.obj <- SummarizedExperiment(geno)
obj <- rqtClass(phenotype=pheno, genotype=geno.obj)
res <- rQTest(obj, method="pca", out.type = "D")
results(res)
}

