% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQTest.R
\name{rQTest}
\alias{rQTest}
\title{This function performs a gene-level test based on combined effect sizes.}
\usage{
rQTest(pheno, geno, perm = 0, STT = 0.2, weight = FALSE,
  cumvar.threshold = 90, out.type = "D", method = "pca")
}
\arguments{
\item{pheno}{Phenotype (vector).}

\item{geno}{Genotype (matrix n by m where n - number of individuals, m - number of genetic variants)/}

\item{perm}{Integer indicating the number of permutations to compute p-values. Default: 0.}

\item{STT}{TODO}

\item{weight}{Logical. TODO}

\item{cumvar.threshold}{Numeric value indicating the explained variance threshold for PCA-like methods. Default: 90.}

\item{out.type}{Character, indicating a type of phenotype. Possible values: D (dichotomous or binary), 
C (continous or qualitative).}

\item{method}{Method used to reduce multicollinerity and account for LD. Default: PCA.}
}
\value{
A list with p-values and corresponding statistics.
}
\description{
TODO
}
\examples{
library(rqt)
data <- read.table(system.file("data/phengen2.dat",package="rqt"), skip=1)
pheno <- data[,1]
geno <- data[, 2:dim(data)[2]]
res <- rQTest(pheno=pheno, geno=geno)
}

