% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqt.R
\name{vcov_rigde}
\alias{vcov_rigde}
\title{Calculates variance-covariance matrix for LASSO/ridge regression.}
\usage{
vcov_rigde(x, y, rmod)
}
\description{
Calculates variance-covariance matrix for LASSO/ridge regression.
}

